extension and reconstruction of the project
