--- src/controller/FakerControllerInterface.ts
@@ -1,3 +1,0 @@
-export interface FakerControllerInterface {
-   createResponse(parsedReq: string): {}
-}
