--- src/models/Colors.ts
@@ -1,6 +1,0 @@
-export const Colors = [
-    'red',
-    '#00ff00',
-    'rgb(0, 0, 255)',
-    'rgba(0, 255, 200, 0.5)'
-]
