--- src/models/Directives.ts
@@ -1,10 +1,0 @@
-export const enum Directives {
-    // '@settings',
-    // '@root',
-    // '@data',
-    // '@use',
-    // '@static',
-    // '@repeat',
-    // '@key'
-    //@format -> dla dat lub wprowadzić pipy
-}
