--- src/models/FakerMethods.ts
@@ -1,15 +1,0 @@
-export const FakerMethods = [
-    'address',
-    'commerce',
-    'company',
-    'database',
-    'finance',
-    'hacker',
-    'helpers',
-    'internet',
-    'lorem',
-    'name',
-    'phone',
-    'random',
-    'system',
-]
