--- src/services/RandomShapeServiceInterface.ts
@@ -1,3 +1,0 @@
-export interface RandomShapeServiceInterface {
-    getRandomShape(shape: string): string
-}
