+++ tsconfig.json
@@ -20,5 +20,5 @@
     }
   },
   "include": ["./src/**/*.ts"],
-  "exclude": ["lib", "build", "node_modules"]
+  "exclude": ["lib", "build", "node_modules", "examples", "img"]
 }
